requirements.txt
tensorflow>=2.11.0
tensorflow-addons>=0.20.0
mtcnn>=0.1.0
opencv-python>=4.5.5
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
